added more parsing methods for the transformation from an eaglemodel back to a xml based .sch file (sheets node is the only missing thing)

+++ src/eaglemodel/Label.java
@@ -198,12 +198,12 @@ public interface Label extends EObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Rot</em>' attribute.
-	 * @see #setRot(double)
+	 * @see #setRot(int)
 	 * @see eaglemodel.EaglemodelPackage#getLabel_Rot()
 	 * @model
 	 * @generated
 	 */
-	double getRot();
+	int getRot();
 
 	/**
 	 * Sets the value of the '{@link eaglemodel.Label#getRot <em>Rot</em>}' attribute.
@@ -213,7 +213,7 @@ public interface Label extends EObject {
 	 * @see #getRot()
 	 * @generated
 	 */
-	void setRot(double value);
+	void setRot(int value);
 
 	/**
 	 * Returns the value of the '<em><b>Xref</b></em>' attribute.
