+++ src/eaglemodel/Pad.java
@@ -199,12 +199,12 @@ public interface Pad extends EObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Rot</em>' attribute.
-	 * @see #setRot(double)
+	 * @see #setRot(int)
 	 * @see eaglemodel.EaglemodelPackage#getPad_Rot()
 	 * @model
 	 * @generated
 	 */
-	double getRot();
+	int getRot();
 
 	/**
 	 * Sets the value of the '{@link eaglemodel.Pad#getRot <em>Rot</em>}' attribute.
@@ -214,7 +214,7 @@ public interface Pad extends EObject {
 	 * @see #getRot()
 	 * @generated
 	 */
-	void setRot(double value);
+	void setRot(int value);
 
 	/**
 	 * Returns the value of the '<em><b>Stop</b></em>' attribute.
