+++ src/eaglemodel/Text.java
@@ -200,12 +200,12 @@ public interface Text extends EObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Rot</em>' attribute.
-	 * @see #setRot(double)
+	 * @see #setRot(int)
 	 * @see eaglemodel.EaglemodelPackage#getText_Rot()
 	 * @model
 	 * @generated
 	 */
-	double getRot();
+	int getRot();
 
 	/**
 	 * Sets the value of the '{@link eaglemodel.Text#getRot <em>Rot</em>}' attribute.
@@ -215,7 +215,7 @@ public interface Text extends EObject {
 	 * @see #getRot()
 	 * @generated
 	 */
-	void setRot(double value);
+	void setRot(int value);
 
 	/**
 	 * Returns the value of the '<em><b>Align</b></em>' attribute.
