+++ src/eaglemodel/impl/PinImpl.java
@@ -206,7 +206,7 @@ public class PinImpl extends MinimalEObjectImpl.Container implements Pin {
 	 * @generated
 	 * @ordered
 	 */
-	protected static final double ROT_EDEFAULT = 0.0;
+	protected static final int ROT_EDEFAULT = 0;
 
 	/**
 	 * The cached value of the '{@link #getRot() <em>Rot</em>}' attribute.
@@ -216,7 +216,7 @@ public class PinImpl extends MinimalEObjectImpl.Container implements Pin {
 	 * @generated
 	 * @ordered
 	 */
-	protected double rot = ROT_EDEFAULT;
+	protected int rot = ROT_EDEFAULT;
 
 	/**
 	 * <!-- begin-user-doc -->
@@ -410,7 +410,7 @@ public class PinImpl extends MinimalEObjectImpl.Container implements Pin {
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public double getRot() {
+	public int getRot() {
 		return rot;
 	}
 
@@ -419,8 +419,8 @@ public class PinImpl extends MinimalEObjectImpl.Container implements Pin {
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setRot(double newRot) {
-		double oldRot = rot;
+	public void setRot(int newRot) {
+		int oldRot = rot;
 		rot = newRot;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, EaglemodelPackage.PIN__ROT, oldRot, rot));
@@ -489,7 +489,7 @@ public class PinImpl extends MinimalEObjectImpl.Container implements Pin {
 				setSwaplevel((Integer)newValue);
 				return;
 			case EaglemodelPackage.PIN__ROT:
-				setRot((Double)newValue);
+				setRot((Integer)newValue);
 				return;
 		}
 		super.eSet(featureID, newValue);
