+++ src/eaglemodel/impl/SMDImpl.java
@@ -184,7 +184,7 @@ public class SMDImpl extends MinimalEObjectImpl.Container implements SMD {
 	 * @generated
 	 * @ordered
 	 */
-	protected static final double ROT_EDEFAULT = 0.0;
+	protected static final int ROT_EDEFAULT = 0;
 
 	/**
 	 * The cached value of the '{@link #getRot() <em>Rot</em>}' attribute.
@@ -194,7 +194,7 @@ public class SMDImpl extends MinimalEObjectImpl.Container implements SMD {
 	 * @generated
 	 * @ordered
 	 */
-	protected double rot = ROT_EDEFAULT;
+	protected int rot = ROT_EDEFAULT;
 
 	/**
 	 * The default value of the '{@link #isStop() <em>Stop</em>}' attribute.
@@ -427,7 +427,7 @@ public class SMDImpl extends MinimalEObjectImpl.Container implements SMD {
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public double getRot() {
+	public int getRot() {
 		return rot;
 	}
 
@@ -436,8 +436,8 @@ public class SMDImpl extends MinimalEObjectImpl.Container implements SMD {
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	public void setRot(double newRot) {
-		double oldRot = rot;
+	public void setRot(int newRot) {
+		int oldRot = rot;
 		rot = newRot;
 		if (eNotificationRequired())
 			eNotify(new ENotificationImpl(this, Notification.SET, EaglemodelPackage.SMD__ROT, oldRot, rot));
@@ -570,7 +570,7 @@ public class SMDImpl extends MinimalEObjectImpl.Container implements SMD {
 				setRoundness((Integer)newValue);
 				return;
 			case EaglemodelPackage.SMD__ROT:
-				setRot((Double)newValue);
+				setRot((Integer)newValue);
 				return;
 			case EaglemodelPackage.SMD__STOP:
 				setStop((Boolean)newValue);
